LayoutTests/imported/w3c/web-platform-tests/html/semantics/popups/popup-top-layer-interactions.tentative-expected.txt

This is a Pop-up API
This is a Pop-up API
This is a Pop-up API
This is a Pop-up API
This is a Fullscreen Element
This is a Fullscreen Element
This is a Fullscreen Element
This is a Fullscreen Element

FAIL A Pop-up API should  close a Pop-up API promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Modal Dialog should  close a Pop-up API promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Fullscreen Element should  close a Pop-up API promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Pop-up API should *not* close a Modal Dialog promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
PASS A Modal Dialog should *not* close a Modal Dialog
FAIL A Fullscreen Element should *not* close a Modal Dialog promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Pop-up API should *not* close a Fullscreen Element promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Modal Dialog should *not* close a Fullscreen Element promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
FAIL A Fullscreen Element should  close a Fullscreen Element promise_test: Unhandled rejection with value: object "SyntaxError: The string did not match the expected pattern."
